+++ src/misc/lv_fs.c
@@ -162,14 +162,14 @@ lv_fs_res_t lv_fs_close(lv_fs_file_t * file_p)
     return res;
 }
 
-static lv_fs_res_t lv_fs_read_cached(lv_fs_file_t * file_p, char * buf, uint32_t btr, uint32_t * br)
+static lv_fs_res_t lv_fs_read_cached(lv_fs_file_t * file_p, char * buf, size_t btr, size_t * br)
 {
     lv_fs_res_t res = LV_FS_RES_OK;
     uint32_t file_position = file_p->cache->file_position;
     uint32_t start = file_p->cache->start;
     uint32_t end = file_p->cache->end;
     char * buffer = file_p->cache->buffer;
-    uint32_t buffer_size = file_p->drv->cache_size;
+    size_t buffer_size = file_p->drv->cache_size;
 
     if(start <= file_position && file_position <= end) {
         /* Data can be read from cache buffer */
@@ -194,7 +194,7 @@ static lv_fs_res_t lv_fs_read_cached(lv_fs_file_t * file_p, char * buf, uint32_t
             file_p->drv->seek_cb(file_p->drv, file_p->file_d, file_p->cache->end + 1,
                                  LV_FS_SEEK_SET);
 
-            uint32_t bytes_read_to_buffer = 0;
+            size_t bytes_read_to_buffer = 0;
             if(btr - buffer_remaining_length > buffer_size) {
                 /*If remaining data chuck is bigger than buffer size, then do not use cache, instead read it directly from FS*/
                 res = file_p->drv->read_cb(file_p->drv, file_p->file_d, (void *)(buf + buffer_remaining_length),
@@ -204,9 +204,9 @@ static lv_fs_res_t lv_fs_read_cached(lv_fs_file_t * file_p, char * buf, uint32_t
                 /*If remaining data chunk is smaller than buffer size, then read into cache buffer*/
                 res = file_p->drv->read_cb(file_p->drv, file_p->file_d, (void *)buffer, buffer_size, &bytes_read_to_buffer);
                 file_p->cache->start = file_p->cache->end + 1;
-                file_p->cache->end = file_p->cache->start + bytes_read_to_buffer - 1;
+                file_p->cache->end = (uint32_t)(file_p->cache->start + bytes_read_to_buffer - 1);
 
-                uint16_t data_chunk_remaining = LV_MIN(btr - buffer_remaining_length, bytes_read_to_buffer);
+                uint16_t data_chunk_remaining = (uint16_t)LV_MIN(btr - buffer_remaining_length, bytes_read_to_buffer);
                 lv_memcpy(buf + buffer_remaining_length, buffer, data_chunk_remaining);
             }
             *br = LV_MIN(buffer_remaining_length + bytes_read_to_buffer, btr);
@@ -229,10 +229,10 @@ static lv_fs_res_t lv_fs_read_cached(lv_fs_file_t * file_p, char * buf, uint32_t
                 buffer = file_p->cache->buffer;
             }
 
-            uint32_t bytes_read_to_buffer = 0;
+            size_t bytes_read_to_buffer = 0;
             res = file_p->drv->read_cb(file_p->drv, file_p->file_d, (void *)buffer, buffer_size, &bytes_read_to_buffer);
             file_p->cache->start = file_position;
-            file_p->cache->end = file_p->cache->start + bytes_read_to_buffer - 1;
+            file_p->cache->end = (uint32_t)(file_p->cache->start + bytes_read_to_buffer - 1);
 
             *br = LV_MIN(btr, bytes_read_to_buffer);
             lv_memcpy(buf, buffer, *br);
@@ -241,19 +241,19 @@ static lv_fs_res_t lv_fs_read_cached(lv_fs_file_t * file_p, char * buf, uint32_t
     }
 
     if(res == LV_FS_RES_OK) {
-        file_p->cache->file_position += *br;
+        file_p->cache->file_position += (uint32_t)*br;
     }
 
     return res;
 }
 
-lv_fs_res_t lv_fs_read(lv_fs_file_t * file_p, void * buf, uint32_t btr, uint32_t * br)
+lv_fs_res_t lv_fs_read(lv_fs_file_t * file_p, void * buf, size_t btr, size_t * br)
 {
     if(br != NULL) *br = 0;
     if(file_p->drv == NULL) return LV_FS_RES_INV_PARAM;
     if(file_p->drv->read_cb == NULL) return LV_FS_RES_NOT_IMP;
 
-    uint32_t br_tmp = 0;
+    size_t br_tmp = 0;
     lv_fs_res_t res;
 
     if(file_p->drv->cache_size) {
@@ -268,7 +268,7 @@ lv_fs_res_t lv_fs_read(lv_fs_file_t * file_p, void * buf, uint32_t btr, uint32_t
     return res;
 }
 
-lv_fs_res_t lv_fs_write(lv_fs_file_t * file_p, const void * buf, uint32_t btw, uint32_t * bw)
+lv_fs_res_t lv_fs_write(lv_fs_file_t * file_p, const void * buf, size_t btw, size_t * bw)
 {
     if(bw != NULL) *bw = 0;
 
@@ -288,17 +288,17 @@ lv_fs_res_t lv_fs_write(lv_fs_file_t * file_p, const void * buf, uint32_t btw, u
         if(res != LV_FS_RES_OK) return res;
     }
 
-    uint32_t bw_tmp = 0;
+    size_t bw_tmp = 0;
     res = file_p->drv->write_cb(file_p->drv, file_p->file_d, buf, btw, &bw_tmp);
     if(bw != NULL) *bw = bw_tmp;
 
     if(file_p->drv->cache_size && res == LV_FS_RES_OK)
-        file_p->cache->file_position += bw_tmp;
+        file_p->cache->file_position += (uint32_t)bw_tmp;
 
     return res;
 }
 
-lv_fs_res_t lv_fs_seek(lv_fs_file_t * file_p, uint32_t pos, lv_fs_whence_t whence)
+lv_fs_res_t lv_fs_seek(lv_fs_file_t * file_p, size_t pos, lv_fs_whence_t whence)
 {
     if(file_p->drv == NULL) {
         return LV_FS_RES_INV_PARAM;
@@ -312,7 +312,7 @@ lv_fs_res_t lv_fs_seek(lv_fs_file_t * file_p, uint32_t pos, lv_fs_whence_t whenc
     if(file_p->drv->cache_size) {
         switch(whence) {
             case LV_FS_SEEK_SET: {
-                    file_p->cache->file_position = pos;
+                    file_p->cache->file_position = (uint32_t)pos;
 
                     /*FS seek if new position is outside cache buffer*/
                     if(file_p->cache->file_position < file_p->cache->start || file_p->cache->file_position > file_p->cache->end) {
@@ -322,7 +322,7 @@ lv_fs_res_t lv_fs_seek(lv_fs_file_t * file_p, uint32_t pos, lv_fs_whence_t whenc
                     break;
                 }
             case LV_FS_SEEK_CUR: {
-                    file_p->cache->file_position += pos;
+                    file_p->cache->file_position += (uint32_t)pos;
 
                     /*FS seek if new position is outside cache buffer*/
                     if(file_p->cache->file_position < file_p->cache->start || file_p->cache->file_position > file_p->cache->end) {
@@ -335,11 +335,11 @@ lv_fs_res_t lv_fs_seek(lv_fs_file_t * file_p, uint32_t pos, lv_fs_whence_t whenc
                     /*Because we don't know the file size, we do a little trick: do a FS seek, then get the new file position from FS*/
                     res = file_p->drv->seek_cb(file_p->drv, file_p->file_d, pos, whence);
                     if(res == LV_FS_RES_OK) {
-                        uint32_t tmp_position;
+                        size_t tmp_position;
                         res = file_p->drv->tell_cb(file_p->drv, file_p->file_d, &tmp_position);
 
                         if(res == LV_FS_RES_OK) {
-                            file_p->cache->file_position = tmp_position;
+                            file_p->cache->file_position = (uint32_t)tmp_position;
                         }
                     }
                     break;
@@ -353,7 +353,7 @@ lv_fs_res_t lv_fs_seek(lv_fs_file_t * file_p, uint32_t pos, lv_fs_whence_t whenc
     return res;
 }
 
-lv_fs_res_t lv_fs_tell(lv_fs_file_t * file_p, uint32_t * pos)
+lv_fs_res_t lv_fs_tell(lv_fs_file_t * file_p, size_t * pos)
 {
     if(file_p->drv == NULL) {
         *pos = 0;
